kinks + infotag safeAnyAscii fixes
